+++ src/test/java/org/scion/jpan/api/DatagramSocketApiServerTest.java
@@ -18,7 +18,6 @@ import static org.junit.jupiter.api.Assertions.*;
 
 import java.io.IOException;
 import java.net.*;
-import java.nio.ByteBuffer;
 import java.nio.file.Paths;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
@@ -36,7 +35,7 @@ import org.scion.jpan.testutil.MockNetwork;
 class DatagramSocketApiServerTest {
 
   @AfterEach
-  public void afterEach() throws IOException {
+  void afterEach() throws IOException {
     MockDaemon.closeDefault();
     MockDNS.clear();
     ScionService.closeDefault();
@@ -165,8 +164,8 @@ class DatagramSocketApiServerTest {
     InetSocketAddress brAddress = IPHelper.toInetSocketAddress(brAddressString);
 
     // prepare packet
-    ScionPacketInspector spi = new ScionPacketInspector();
-    spi.read(ByteBuffer.wrap(ExamplePacket.PACKET_BYTES_SERVER_E2E_PING));
+    ScionPacketInspector spi =
+        ScionPacketInspector.wrapReadOnly(ExamplePacket.PACKET_BYTES_SERVER_E2E_PING);
     spi.getScionHeader().setSrcHostAddress(scionSrcBytes);
     InetAddress scionSrcIP = spi.getScionHeader().getSrcHostAddress();
     int scionSrcPort = spi.getOverlayHeaderUdp().getSrcPort();
@@ -213,8 +212,8 @@ class DatagramSocketApiServerTest {
     InetSocketAddress underLaySrc =
         new InetSocketAddress(InetAddress.getByAddress(new byte[] {10, 0, 123, 123}), 12345);
 
-    ScionPacketInspector spi = new ScionPacketInspector();
-    spi.read(ByteBuffer.wrap(ExamplePacket.PACKET_BYTES_SERVER_E2E_PING));
+    ScionPacketInspector spi =
+        ScionPacketInspector.wrapReadOnly(ExamplePacket.PACKET_BYTES_SERVER_E2E_PING);
     spi.getPathHeaderScion().reset(); // set path to []
     InetAddress scionSrcIP = spi.getScionHeader().getSrcHostAddress();
     int scionSrcPort = spi.getOverlayHeaderUdp().getSrcPort();